+++ app/src/main/java/com/example/mediaplayer/util/Constants.kt
@@ -16,8 +16,10 @@ object Constants {
     const val READ_STORAGE = Manifest.permission.READ_EXTERNAL_STORAGE
     const val WRITE_STORAGE = Manifest.permission.WRITE_EXTERNAL_STORAGE
 
-    const val STORE_PERMISSION_MSG = ""
+    // Music Service
+    const val MEDIA_ROOT_ID = "Kylentt_root_id"
 
+    // Notification
     const val NOTIFICATION_CHANNEL_ID = "MediaPlayer"
     const val NOTIFICATION_ID = 301
 }