+++ Changelog.md
@@ -1,3 +1,8 @@
+# 2022/01/30
+- added MusicSource class for converting fetched song metadata to MediaMetadataCompat
+- setting up NotificationManager class for handling Notification by MusicService Class
+- added NotificationListener class to Listen for NotificationManager changes / callbacks
+
 # 2022/01/29
 - yet Another fix to Player Inconsistency `will migrate the player to a Service`
 - yet Another fix to problem when song is deleted in Storage `will migrate using Room & Datastore`